# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API EliteRT",
        "version": "1.0.0"
    },
    "paths": {
        "/api/detail-penghuni": {
            "get": {
                "tags": [
                    "Detail Penghuni Rumah"
                ],
                "summary": "Menampilkan semua data detail penghuni rumah",
                "operationId": "2c325e0f9f5d8eebea39ab657be1a986",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Detail Penghuni Rumah"
                ],
                "summary": "Menambahkan data detail penghuni rumah",
                "operationId": "abd904308d7d63ea64c6d5278eebceb8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_rumah",
                                    "id_penghuni"
                                ],
                                "properties": {
                                    "id_rumah": {
                                        "type": "integer"
                                    },
                                    "id_penghuni": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/detail-penghuni/{id}": {
            "delete": {
                "tags": [
                    "Detail Penghuni Rumah"
                ],
                "summary": "Menghapus data detail penghuni rumah",
                "operationId": "4e1aac354149bdbecc42bfdb32d6cb6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}